Address Claude and CodeRabbit review feedback
- Extract shared managed-runtime infra into runtime.py so conversion.py no
  longer imports private symbols from whitebox.py (both now import the shared
  module) [Claude: coupling]
- Enforce the conversion timeout with a watchdog Timer that kills a hung
  subprocess even when it emits no output; the previous process.wait(timeout)
  was dead code because the stdout read loop had already drained on exit
  [Claude: dead timeout]
- Terminate the subprocess in a finally block to avoid leaking it when an
  exception is raised mid-stream [CodeRabbit]
- Replace the production assert on process.stdout with an explicit raise
  [Claude: assert in prod]
- Drop the extra COUNT(*) scan; take the feature count from COPY's return
  value instead, halving I/O on large files [Claude: double scan]
- Match CRS-annotated GEOMETRY types by prefix in the vector script
- Reorder the empty-string guard before expanduser() in _validate_paths
  [Claude]
- Sort finished jobs by created_at before eviction so the genuinely oldest
  are dropped [CodeRabbit] + add eviction unit tests
- Auto-scroll the conversion job log and use index-only React keys
  [Claude/CodeRabbit]
- Document row_group_size/compression defaults and validation via JSDoc
  [CodeRabbit]
